==> 02_time_res/17_6_1/00_program/02_split.py <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
from scipy.stats import norm
from sklearn.preprocessing import StandardScaler
from scipy import stats
import warnings
import os
import math

def get_index(date_t,date):
	rows=df_CSAD.shape[0]
	for i in range(1,rows):
		if date_t.iloc[i][1]==date:
			return i
			
path=r'D:\paper_data\01_CCK\02_processed_data\CSAD_Whole.csv'
df_CSAD = pd.read_csv(path,encoding='gb18030')
j=get_index(df_CSAD,'2017/6/1   ')
i=get_index(df_CSAD,'2020/1/23  ')
print(i)

CSAD_Pre=df_CSAD.iloc[j:i+1,1:]
CSAD_Post=df_CSAD.iloc[i+1:,1:]
CSAD_Whole=df_CSAD.iloc[j:,1:]

CSAD_Pre.to_csv(r'D:\paper_data\02_time_res\\17_6_1\02_processed_data\CSAD_Pre.csv')
CSAD_Post.to_csv(r'D:\paper_data\02_time_res\\17_6_1\02_processed_data\CSAD_Post.csv')
CSAD_Whole.to_csv(r'D:\paper_data\02_time_res\\17_6_1\02_processed_data\CSAD_Whole.csv')

path_2=r'D:\paper_data\00_indices\02_processed_data\SWI_indices_Whole.csv'
df_ind = pd.read_csv(path_2,encoding='gb18030')

ind_Pre=df_ind.iloc[j:i+1,1:]
ind_Post=df_ind.iloc[i+1:,1:]
ind_Whole=df_ind.iloc[j:,1:]
ind_Pre.to_csv(r'D:\paper_data\00_indices\03_time_res\\17_6_1\SWI_indices_Pre.csv',encoding='gb18030')
ind_Post.to_csv(r'D:\paper_data\00_indices\03_time_res\\17_6_1\SWI_indices_Post.csv',encoding='gb18030')
ind_Whole.to_csv(r'D:\paper_data\00_indices\03_time_res\\17_6_1\SWI_indices_Whole.csv',encoding='gb18030')
